Rename accelerate_sections -> accelerate, remove accelerate_part
